init item.go files
